Order Controller

+++ src/main/java/com/pro_sb_ecommerce/cart/dto/CartItemResponse.java
@@ -3,13 +3,15 @@ package com.pro_sb_ecommerce.cart.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class CartItemResponse {
 
     private Long productId;
     private String productName;
-    private double priceAtTime;
+    private BigDecimal priceAtTime;
     private int quantity;
-    private double totalPrice;
+    private BigDecimal totalPrice;
 }
